openharmony-schedule - 44

+++ openharmony/arkcompiler_ets_runtime/ecmascript/compiler/assembler/aarch64/assembler_aarch64.h
@@ -339,7 +339,7 @@ public:
     void Br(const Register &rn);
     void Blr(const Register &rn);
     void Bl(Label *label);
-    void Bl(intptr_t imm);
+    void Bl(int32_t imm);
     void Cbz(const Register &rt, int32_t imm);
     void Cbz(const Register &rt, Label *label);
     void Cbnz(const Register &rt, int32_t imm);
